refactor: rename queries and util modules for consistency
