+++ framework/kibocommerce/api/endpoints/cart/get-cart.ts
@@ -2,7 +2,7 @@ import CookieHandler from '@framework/api/utils/cookie-handler'
 import { normalizeCart } from '@framework/lib/normalize'
 import { Cart } from '@framework/schema'
 import type { CartEndpoint } from '.'
-import { getCartQuery } from '../../queries/getCartQuery'
+import { getCartQuery } from '../../queries/get-cart-query'
 
 const getCart: CartEndpoint['handlers']['getCart'] = async ({
   req,
