+++ framework/kibocommerce/api/endpoints/cart/update-item.ts
@@ -1,6 +1,6 @@
 import { normalizeCart } from '../../../lib/normalize'
 import type { CartEndpoint } from '.'
-import { getCartQuery } from '@framework/api/queries/getCartQuery'
+import { getCartQuery } from '@framework/api/queries/get-cart-query'
 import updateCartItemQuantityMutation from '@framework/api/mutations/updateCartItemQuantity-mutation'
 
 const updateItem: CartEndpoint['handlers']['updateItem'] = async ({
